pi/pi_requirements.txt
xmljson
pyserial
requests
pycrypto
pyyaml
